Fixed types

+++ src/nes.rs
@@ -104,8 +104,8 @@ impl Nes {
             APP_NAME.to_owned()
         };
 
-        let width = (self.config.scale * WINDOW_WIDTH as f32) as u32;
-        let height = (self.config.scale * WINDOW_HEIGHT as f32) as u32;
+        let width = (self.config.scale * WINDOW_WIDTH as f32) as i32;
+        let height = (self.config.scale * WINDOW_HEIGHT as f32) as i32;
         let mut engine = PixEngine::builder();
         engine
             .with_dimensions(width, height)
@@ -136,7 +136,7 @@ impl AppState for Nes {
     fn on_start(&mut self, s: &mut PixState) -> PixResult<()> {
         let main_window = WindowBuilder::new(s.width(), s.height())
             .with_id(s.window_id())
-            .create_texture(PixelFormat::Rgb, RENDER_WIDTH, RENDER_HEIGHT)
+            .create_texture(PixelFormat::Rgb, RENDER_WIDTH as i32, RENDER_HEIGHT as i32)
             .clip([0, 8, RENDER_WIDTH as i32, RENDER_HEIGHT as i32 - 8])
             .build(s)?;
         self.windows.push(main_window);
